Move mocked data into individual hook categories
