+++ packages/commerce-sdk-react/jest.config.js
@@ -6,7 +6,6 @@
  */
 
 const base = require('internal-lib-build/configs/jest/jest.config')
-// const path = require('path')
 
 module.exports = {
     ...base,
@@ -18,5 +17,5 @@ module.exports = {
             statements: 0
         }
     },
-    collectCoverageFrom: ['src/**/*.{js,jsx,ts,tsx}']
+    collectCoverageFrom: ['src/**/*.{js,jsx,ts,tsx}', '!src/**/mock/*']
 }
