+++ packages/commerce-sdk-react/src/hooks/ShopperProducts/mock/index.ts
